+++ flutter/lib/src/sentry_flutter.dart
@@ -16,6 +16,7 @@ import 'flutter_exception_type_identifier.dart';
 import 'frame_callback_handler.dart';
 import 'integrations/connectivity/connectivity_integration.dart';
 import 'integrations/integrations.dart';
+import 'integrations/native_app_start_handler.dart';
 import 'integrations/screenshot_integration.dart';
 import 'native/factory.dart';
 import 'native/native_scope_observer.dart';
@@ -194,10 +195,12 @@ mixin SentryFlutter {
     integrations.add(LoadReleaseIntegration());
 
     if (native != null) {
-      integrations.add(NativeAppStartIntegration(
-        native,
-        DefaultFrameCallbackHandler(),
-      ));
+      integrations.add(
+        NativeAppStartIntegration(
+          DefaultFrameCallbackHandler(),
+          NativeAppStartHandler(native),
+        ),
+      );
     }
     return integrations;
   }
@@ -213,9 +216,15 @@ mixin SentryFlutter {
   }
 
   /// Manually set when your app finished startup. Make sure to set
-  /// [SentryFlutterOptions.autoAppStart] to false on init.
+  /// [SentryFlutterOptions.autoAppStart] to false on init. The timeout duration
+  /// for this to work is 10 seconds.
   static void setAppStartEnd(DateTime appStartEnd) {
-    NativeAppStartIntegration.appStartEnd = appStartEnd;
+    // ignore: invalid_use_of_internal_member
+    final integrations = Sentry.currentHub.options.integrations
+        .whereType<NativeAppStartIntegration>();
+    for (final integration in integrations) {
+      integration.appStartEnd = appStartEnd;
+    }
   }
 
   static void _setSdk(SentryFlutterOptions options) {